requirements.txt
requests
pandas
python-dotenv
tabulate
openai
streamlit